Reduce JS module file count
