+++ www/assets/js/templates/cookbook.js
@@ -1,3 +1,2 @@
-import { component } from "../utils/load.js";
 
-component("../components/filter.js", "#cookbook");
+component("./components/filter.js", "#cookbook");
